refactor: domain package 추가
